feat: project owner only can delete or update project and member can update or delete feature

+++ src/controllers/featureControllers.ts
@@ -1,15 +1,26 @@
 import { Request, Response, NextFunction } from 'express';
 import Feature from '../models/featureModels';
+import { AppRequest } from '../types/custom/user';
+import Project from '../models/projectModels';
 
-const createFeature = async (req: Request, res: Response, next: NextFunction)=>{
-    const {title,description,color} = req.body
+const createFeature = async (req: AppRequest, res: Response, next: NextFunction) => {
+    const { title, description, color } = req.body
     const { id } = req.params;
-    try{
-        await Feature.create({projectId:id,title,description,color})
+    const userId = req.appUser?._id
+
+    try {
+        const proj = await Project.findOne({ _id: id })
+        if (!proj) {
+            throw new Error('Project not found')
+        }
+        if (proj?.projectOwner.toString() !== userId) {
+            throw new Error('You are not authorized to create a feature for this project')
+        }
+        await Feature.create({ projectId: id, title, description, color })
         res.status(201).json({
-            message:'Feature added successfully'
+            message: 'Feature added successfully'
         })
-    }catch(err){
+    } catch (err) {
         if (err instanceof Error) {
             res.status(500).json({ message: err.message });
         } else {
@@ -18,18 +29,29 @@ const createFeature = async (req: Request, res: Response, next: NextFunction)=>{
     }
 }
 
-const updateFeature = async (req: Request, res: Response, next: NextFunction)=>{
+const updateFeature = async (req: AppRequest, res: Response, next: NextFunction) => {
     const { id } = req.params;
     const feature = req.body
-    try{
-        const feat = await Feature.updateOne({_id:id}, feature)
-        if(feat.matchedCount===0){
+    const userId = req.appUser?._id
+    try {
+        const feat = await Feature.findOne({ _id: id })
+        if (!feat) {
+            throw new Error('Feature not found')
+        }
+        const project = await Project.findOne({ _id: feat?.projectId })
+        const isUserMemberOfProject = project?.members.find((user)=>user.userId.toString() == userId)
+        const isUserOwnerOfProject = project?.projectOwner.toString() == userId
+        if (!isUserOwnerOfProject || !isUserMemberOfProject) {
+            throw new Error('You are not authorized to update this feature')
+        }
+        const updatedfeature = await Feature.updateOne({ _id: id }, feature)
+        if (updatedfeature.matchedCount === 0) {
             throw new Error('This feature not found')
         }
         res.status(201).json({
-            message:'Feature updated successfully'
+            message: 'Feature updated successfully'
         })
-    }catch(err){
+    } catch (err) {
         if (err instanceof Error) {
             res.status(500).json({ message: err.message });
         } else {
@@ -38,15 +60,15 @@ const updateFeature = async (req: Request, res: Response, next: NextFunction)=>{
     }
 }
 
-const getAllFeature = async (req: Request, res: Response, next: NextFunction)=>{
+const getAllFeature = async (req: Request, res: Response, next: NextFunction) => {
     const { id } = req.params;
-    try{
-        const Features = await Feature.find({projectId:id})
+    try {
+        const Features = await Feature.find({ projectId: id })
         res.status(201).json({
-            message:'Features get successfully',
-			data: Features,
+            message: 'Features get successfully',
+            data: Features,
         })
-    }catch(err){
+    } catch (err) {
         if (err instanceof Error) {
             res.status(500).json({ message: err.message });
         } else {
@@ -55,17 +77,28 @@ const getAllFeature = async (req: Request, res: Response, next: NextFunction)=>{
     }
 }
 
-const deleteFeature = async (req: Request, res: Response, next: NextFunction)=>{
+const deleteFeature = async (req: AppRequest, res: Response, next: NextFunction) => {
     const { id } = req.params;
-    try{
-        let feat = await Feature.findOneAndDelete({_id:id})
-        if(!feat){
+    const userId: any = req.appUser?._id
+    try {
+        const feat = await Feature.findOne({ _id: id })
+        if (!feat) {
+            throw new Error('Feature not found')
+        }
+        const project = await Project.findOne({ _id: feat?.projectId })
+        const isUserMemberOfProject = project?.members.find((user)=>user.userId.toString() == userId)
+        const isUserOwnerOfProject = project?.projectOwner.toString() == userId
+        if (!isUserOwnerOfProject || !isUserMemberOfProject) {
+            throw new Error('You are not authorized to delete this feature')
+        }
+        let deletedFeature = await Feature.findOneAndDelete({ _id: id })
+        if (!deletedFeature) {
             throw new Error('This feature not found')
         }
         res.status(201).json({
-            message:'Project deleted successfully',
+            message: 'Project deleted successfully',
         })
-    }catch(err){
+    } catch (err) {
         if (err instanceof Error) {
             res.status(500).json({ message: err.message });
         } else {
@@ -74,4 +107,4 @@ const deleteFeature = async (req: Request, res: Response, next: NextFunction)=>{
     }
 }
 
-export { createFeature,updateFeature,getAllFeature,deleteFeature };
+export { createFeature, updateFeature, getAllFeature, deleteFeature };
